Switch to .F90 suffixes for Fortran source files
The .F90 suffix is assumed by the GNU, Intel, PGI, NVHPC, Cray, and NAG
compilers to signify free-format source files that require pre-processing.
The change in file suffix obviates the need for the -ffree-form compiler flag,
which is GNU-specific.

This commit also removes the GNU-specific -Wall flag, resulting in a top-level
Makefile that should be compiler independent.